Added macos environment
added macos environment
